Added groups to one file
