--- core/tests/protractor/subscriptionsFlow.js
@@ -1,106 +1,0 @@
-// Copyright 2018 The Oppia Authors. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS-IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-/**
- * @fileoverview End-to-end tests for the subscriptions functionality on desktop
- * and mobile.
- */
-
-var general = require('../protractor_utils/general.js');
-var users = require('../protractor_utils/users.js');
-var waitFor = require('../protractor_utils/waitFor.js');
-
-var AdminPage = require('../protractor_utils/AdminPage.js');
-var CreatorDashboardPage =
-  require('../protractor_utils/CreatorDashboardPage.js');
-var PreferencesPage = require('../protractor_utils/PreferencesPage.js');
-var SubscriptionDashboardPage =
-  require('../protractor_utils/SubscriptionDashboardPage.js');
-
-describe('Subscriptions functionality', function() {
-  var creatorDashboardPage = null;
-  var preferencesPage = null;
-  var subscriptionDashboardPage = null;
-
-  beforeEach(function() {
-    creatorDashboardPage = new CreatorDashboardPage.CreatorDashboardPage();
-    preferencesPage = new PreferencesPage.PreferencesPage();
-    subscriptionDashboardPage = (
-      new SubscriptionDashboardPage.SubscriptionDashboardPage());
-  });
-
-  it('handle subscriptions to creators correctly', function() {
-    // Create two creators.
-    users.createUser('creator1Id@subscriptions.com', 'creator1Idsubscriptions');
-    users.createUser('creator2Id@subscriptions.com', 'creator2Idsubscriptions');
-
-    // Create a learner who subscribes to both the creators.
-    users.createUser('learner1@subscriptions.com', 'learner1subscriptions');
-    users.login('learner1@subscriptions.com');
-    subscriptionDashboardPage.navigateToUserSubscriptionPage(
-      'creator1Idsubscriptions');
-    subscriptionDashboardPage.navigateToSubscriptionButton();
-    subscriptionDashboardPage.navigateToUserSubscriptionPage(
-      'creator2Idsubscriptions');
-    subscriptionDashboardPage.navigateToSubscriptionButton();
-    preferencesPage.get();
-    preferencesPage.expectDisplayedFirstSubscriptionToBe('creator...');
-    preferencesPage.expectDisplayedLastSubscriptionToBe('creator...');
-    users.logout();
-
-    // Create a learner who subscribes to creator1Id and unsubscribes from the
-    // creator2Id.
-    users.createUser('learner2@subscriptions.com', 'learner2subscriptions');
-    users.login('learner2@subscriptions.com');
-    subscriptionDashboardPage.navigateToUserSubscriptionPage(
-      'creator1Idsubscriptions');
-    subscriptionDashboardPage.navigateToSubscriptionButton();
-    subscriptionDashboardPage.navigateToUserSubscriptionPage(
-      'creator2Idsubscriptions');
-
-    // Subscribe and then unsubscribe from the same user.
-    subscriptionDashboardPage.navigateToSubscriptionButton();
-    subscriptionDashboardPage.navigateToSubscriptionButton();
-    preferencesPage.get();
-    preferencesPage.expectSubscriptionCountToEqual(1);
-    preferencesPage.expectDisplayedFirstSubscriptionToBe('creator...');
-    users.logout();
-
-    // Verify there are 2 subscribers.
-    users.login('creator1Id@subscriptions.com');
-    // Go to the creator_dashboard.
-    creatorDashboardPage.get();
-    creatorDashboardPage.clickCreateNewExplorationButton();
-    creatorDashboardPage.get();
-    creatorDashboardPage.navigateToSubscriptionDashboard();
-    subscriptionDashboardPage.expectSubscriptionFirstNameToMatch('learner...');
-    subscriptionDashboardPage.expectSubscriptionLastNameToMatch('learner...');
-    users.logout();
-
-    // Verify there are 1 subscriber.
-    users.login('creator2Id@subscriptions.com');
-    // Go to the creator_dashboard.
-    creatorDashboardPage.get();
-    creatorDashboardPage.clickCreateNewExplorationButton();
-    creatorDashboardPage.get();
-    creatorDashboardPage.navigateToSubscriptionDashboard();
-    subscriptionDashboardPage.expectSubscriptionCountToEqual(1);
-    subscriptionDashboardPage.expectSubscriptionLastNameToMatch('learner...');
-    users.logout();
-  });
-
-  afterEach(function() {
-    general.checkForConsoleErrors([]);
-  });
-});
